# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: GProA - Punto de Venta Distribuidor IA
  description: API REST para sistema de punto de venta, inventario y gestión de bodega
  version: 1.0.0
  contact:
    name: GProA Technology
    email: support@gproa.com
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo
...
paths:
...
  /api/inventory/products:
    get:
      tags:
        - Inventory
      summary: Listar productos
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        200:
          description: Lista de productos
    post:
      tags:
        - Inventory
      summary: Crear producto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        201:
          description: Producto creado
...
components:
  schemas:
    Product:
      type: object
      required:
        - sku
        - name
        - category
        - price
        - cost
        - minStock
      properties:
        sku:
          type: string
          example: LAC-001
        name:
          type: string
          example: Leche Entera 1L
        category:
          type: string
          example: Lácteos
        price:
          type: number
          example: 28.00
        cost:
          type: number
          example: 22.00
        stock:
          type: number
          example: 150
        minStock:
          type: number
          example: 50
...